MF-031 Extract navigationService from graphService
